Move SeedUserData to the domain layer
